Week 4 - Midterm Certification/requirements.txt
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
async-lru==2.0.4
asyncer==0.0.2
attrs==23.2.0
babel==2.14.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
bidict==0.23.1
bleach==6.1.0
bs4==0.0.2
build==1.1.1
bzip2==1.0.8
ca-certificates==2023.12.12
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chainlit==0.7.700
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.7
cohere==4.37
coloredlogs==15.0.1
comm==0.2.1
contourpy==1.2.0
curl-cffi==0.6.1
cycler==0.12.1
dataclasses-json==0.5.14
datasets==2.18.0
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
deprecated==1.2.14
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
docker-pycreds==0.4.0
duckduckgo-search==4.5.0
executing==2.0.1
faiss-cpu==1.8.0
fastapi==0.100.1
fastapi-socketio==0.0.10
fastavro==1.9.4
fastjsonschema==2.19.1
filelock==3.13.1
filetype==1.2.0
flatbuffers==24.3.7
fonttools==4.49.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.2.0
gitdb==4.0.11
gitpython==3.1.42
google-auth==2.28.2
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.62.0
h11==0.14.0
httpcore==0.17.3
httptools==0.6.1
httpx==0.24.1
huggingface-hub==0.20.3
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.3
ipykernel==6.29.2
ipython==8.22.1
ipywidgets==8.1.2
isoduration==20.11.0
jedi==0.19.1
jinja2==3.1.3
joblib==1.3.2
json5==0.9.17
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-client==8.6.0
jupyter-console==6.6.3
jupyter-core==5.7.1
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter-server==2.12.5
jupyter-server-terminals==0.5.2
jupyterlab==4.1.2
jupyterlab-pygments==0.3.0
jupyterlab-server==2.25.3
jupyterlab-widgets==3.0.10
kiwisolver==1.4.5
kubernetes==29.0.0
langchain==0.1.11
langchain-community==0.0.27
langchain-core==0.1.30
langchain-openai==0.0.8
langchain-pinecone==0.0.3
langchain-text-splitters==0.0.1
langchainhub==0.1.15
langsmith==0.1.21
lazify==0.4.0
ld_impl_linux-64==2.38
libffi==3.4.4
libgcc-ng==11.2.0
libgomp==11.2.0
libstdcxx-ng==11.2.0
libuuid==1.41.5
llama-index==0.10.18
llama-index-agent-openai==0.1.5
llama-index-callbacks-wandb==0.1.2
llama-index-cli==0.1.8
llama-index-core==0.10.18.post1
llama-index-embeddings-openai==0.1.6
llama-index-indices-managed-llama-cloud==0.1.3
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.7
llama-index-multi-modal-llms-openai==0.1.4
llama-index-program-openai==0.1.4
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.9
llama-index-readers-llama-parse==0.1.3
llama-index-readers-wikipedia==0.1.3
llama-index-vector-stores-chroma==0.1.6
llama-parse==0.3.8
llamaindex-py-client==0.1.13
lxml==5.1.0
markupsafe==2.1.5
marshmallow==3.20.2
matplotlib==3.8.3
matplotlib-inline==0.1.6
mistune==3.0.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
nbclient==0.9.0
nbconvert==7.16.1
nbformat==5.9.2
ncurses==6.4==h6a678d5_0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
notebook==7.1.0
notebook-shim==0.2.4
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
onnxruntime==1.17.1
openai==1.13.3
openssl==3.0.13==h7f8727e_0
opentelemetry-api==1.23.0
opentelemetry-exporter-otlp==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-exporter-otlp-proto-http==1.23.0
opentelemetry-instrumentation==0.44b0
opentelemetry-instrumentation-asgi==0.44b0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-proto==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
opentelemetry-util-http==0.44b0
orjson==3.9.15
overrides==7.7.0
packaging==23.2
pandas==2.2.1
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
pinecone-client==3.1.0
pip==23.3.1
platformdirs==4.2.0
plotly==5.19.0
posthog==3.5.0
prometheus-client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pulsar-client==3.4.0
pure-eval==0.2.2
pyarrow==15.0.0
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.11
pydantic-core==2.16.3
pygments==2.17.2
pyjwt==2.8.0
pymupdf==1.23.26
pymupdfb==1.23.22
pyparsing==3.1.1
pypdf==4.1.0
pypdf2==3.0.1
pypika==0.48.9
pyproject-hooks==1.0.0
pysbd==0.3.4
python==3.11.7
python-dateutil==2.8.2
python-dotenv==1.0.1
python-engineio==4.9.0
python-graphql-client==0.4.3
python-json-logger==2.0.7
python-multipart==0.0.6
python-socketio==5.11.1
pytz==2024.1
pyyaml==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
qtpy==2.4.1
ragas==0.1.3
readline==8.2
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.0
rsa==4.9
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
send2trash==1.8.2
sentence-transformers==2.2.2
sentencepiece==0.2.0
sentry-sdk==1.40.5
setproctitle==1.3.3
setuptools==68.2.2
simple-websocket==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
sqlalchemy==2.0.28
sqlite==3.41.2
stack-data==0.6.3
starlette==0.27.0
striprtf==0.0.26
sympy==1.12
syncer==2.0.3
tenacity==8.2.3
terminado==0.18.0
threadpoolctl==3.3.0
tiktoken==0.6.0
tinycss2==1.2.1
tk==8.6.12
tokenizers==0.15.2
tomli==2.0.1
torch==2.2.1
torchvision==0.17.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
transformers==4.38.2
triton==2.2.0
typer==0.9.0
types-python-dateutil==2.8.19.20240106
types-requests==2.31.0.20240218
typing-extensions==4.9.0
typing-inspect==0.9.0
tzdata==2024.1
uptrace==1.22.0
uri-template==1.3.0
urllib3==2.2.1
uvicorn==0.23.2
uvloop==0.19.0
wandb==0.16.4
watchfiles==0.20.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
wheel==0.41.2
widgetsnbextension==4.0.10
wikipedia==1.4.0
wrapt==1.16.0
wsproto==1.2.0
xxhash==3.4.1
xz==5.4.5
yarl==1.9.4
zipp==3.17.0
zlib==1.2.13